feat:마이페이지 -작성한 리뷰,좋아요한 전통주 리스트 추가

+++ sooltoryteller/src/main/java/com/sooltoryteller/domain/LiqLikeVO.java
@@ -17,7 +17,7 @@ public class LiqLikeVO {
     private String memberId;
 
     // 좋아요 
-    private int liqId;
+    private Long liqId;
 
     // 등록일시 
     private Date regdate;
